Read app/requirements.txt
paho-mqtt
flask
flask_sqlalchemy
azure-iot-device
azure-iot-hub
azure-eventhub
requests
